+++ freediscovery/server/tests/test_search.py
@@ -10,7 +10,8 @@ import pytest
 import json
 import itertools
 from unittest import SkipTest
-from numpy.testing import assert_equal, assert_almost_equal
+import numpy as np
+from numpy.testing import assert_equal, assert_array_less
 
 from .. import fd_app
 from ...utils import _silent, dict2type, sdict_keys
@@ -18,27 +19,46 @@ from ...ingestion import DocumentIndex
 from ...exceptions import OptionalDependencyMissing
 from ...tests.run_suite import check_cache
 
-from .base import parse_res, V01, app, app_notest, get_features_cached, get_features_lsi
+from .base import (parse_res, V01, app, app_notest, get_features_cached,
+                   get_features_lsi, get_features_lsi_cached)
 
 
-@pytest.mark.parametrize("method", ['regular', 'semantic'])
-def test_search(app, method):
+@pytest.mark.parametrize("method, min_score", [('regular', -1),
+                                               ('semantic', -1),
+                                               ('semantic', 0.5)])
+def test_search(app, method, min_score):
 
-    if method == 'regular':
-        dsid, lsi_id, _ = get_features_lsi(app, hashed=False)
+    if method == 'semantic':
+        dsid, lsi_id, _ = get_features_lsi_cached(app, hashed=False)
         parent_id = lsi_id
-    elif method == 'semantic':
+    elif method == 'regular':
         dsid, _ = get_features_cached(app, hashed=False)
         lsi_id = None
         parent_id = dsid
+    query = """The American National Standards Institute sells ANSI standards, and also
+    ISO (international) standards.  Their sales office is at 1-212-642-4900,
+    mailing address is 1430 Broadway, NY NY 10018.  It helps if you have the
+    complete name and number.
+    """
+    query_file_path = "02256.txt"
 
-    method = V01 + "/search/"
-    res = app.post(method, json=dict(parent_id=parent_id, query="so that I can reserve a room"))
+    res = app.post(V01 + "/search/", json=dict(parent_id=parent_id,
+                                               min_score=min_score,
+                                               query=query,
+                                               ))
     assert res.status_code == 200
     data = parse_res(res)
     assert sorted(data.keys()) == ['data']
-    for row in data['data']:
-        if method == 'semantic':
-            assert sorted(row.keys()) == sorted(['score', 'internal_id'])
-        elif method == 'regular':
-            assert sorted(row.keys()) == sorted(['score', 'internal_id', 'document_id'])
+    data = data['data']
+    for row in data:
+        assert dict2type(row) == {'score': 'float',
+                                  'internal_id': 'int',
+                                  'document_id': 'int'}
+    scores = np.array([row['score'] for row in data])
+    assert_equal(np.diff(scores) <= 0, True)
+    assert_array_less(min_score, scores)
+
+    res = app.post(V01 + "/feature-extraction/{}/id-mapping".format(dsid), 
+                   json={'data': [data[0]]})
+    res = parse_res(res)
+    assert res['data'][0]['file_path'] == query_file_path